# speedup persp-maybe-kill-buffer