--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -61,6 +61,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- `persp-maybe-kill-buffer`: kill `persp--make-ignore-buffer-rx` temporary buffers (aka `ido-ignore-buffers`) skipping checks.
 - `persp-kill`: implement `persp-feature-flag-prevent-killing-last-buffer-in-perspective`.
 - `persp-mode`: implement `persp-feature-flag-prevent-killing-last-buffer-in-perspective`.
 - `perspective-map`: Add binding `C-x x B` to call `persp-switch-to-scratch-buffer`.