--- a/test/test-perspective.el
+++ b/test/test-perspective.el
@@ -26,6 +26,9 @@
 (customize-set-variable 'persp-feature-flag-prevent-killing-last-buffer-in-perspective t)
 (customize-set-variable 'persp-feature-flag-directly-kill-ido-ignore-buffers nil)
 
+;; Set `ido-ignore-buffers' to detect temporary buffers.
+(customize-set-variable 'ido-ignore-buffers '("\\` "))
+
 (defun persp-test-interesting-buffer? (buf)
   "Return t if BUF is a non-temporary buffer (i.e., lacks
 tell-tale leading '*' characters)."