@mehw 2021-12-15
> Unfortunately, I'm running into trouble with some basics. In particular, commit [4b0ef93](https://github.com/nex3/perspective-el/commit/4b0ef93dd34e4a595f412a83d95a1583ad70eb4b) introduces a problem: `find-file` instantly errors out.
> 
> It's reproducible on my system just by doing this:
> 
> ```
> emacs -Q -l ~/Code/perspective-el/perspective.el --eval '(persp-mode)' --eval '(toggle-debug-on-error)' --eval '(find-file "~/Code/perspective-el/perspective.el")'
> ```

@gcv: It's quite odd... thanks for bringing that out. I'm looking into it... I didn't manage to sort out the problem yet.