@gcv 2022-04-22
@mehw: Do you think you'll have time to look into this?

If not, I've been using `persp-feature-flag-prevent-killing-last-buffer-in-perspective` and it seems to work quite well. I can flip it to on by default, and quietly get rid of it in a few weeks if no one complains.