@mehw 2022-04-22
@gcv: Yes, sorry, a lot has happened... I was optimistic that I could get back on track uninterrupted, but sometimes problems await us just around the corner...

I look into it.  I have to rebase the PR and make the point about the [error you reported](https://github.com/nex3/perspective-el/pull/171#issuecomment-992141341):
```
emacs -Q -l ~/Code/perspective-el/perspective.el --eval '(persp-mode)' --eval '(toggle-debug-on-error)' --eval '(find-file "~/Code/perspective-el/perspective.el")'
```

I was able to address the problem, but still it's not clear to me the real source of it.