@mehw 2022-05-04
@gcv: Hi, I rebased the patch set on master, all tests passed.

About the error triggered by the below command, `vc-git--run-command-string` is involved, and apparently it's due to the value of `current-buffer`.  I believe to be on a good lead, though.  Thanks for the patience ;)
```
emacs -Q -l ~/Code/perspective-el/perspective.el --eval '(persp-mode)' --eval '(toggle-debug-on-error)' --eval '(find-file "~/Code/perspective-el/perspective.el")'
```