@gcv 2022-05-19
Does the `current-buffer` problem show up because some code involving `vc-git-*` uses `kill-buffer` internally on temporary buffers — which triggers all persp-related checks and special handling?