@gcv 2022-05-19
`kill-buffer` should always and immediately succeed on temporary buffers. That was the point of the workaround I put in: https://github.com/nex3/perspective-el/blob/4e38680793585a907ae46b148697030c2b552a00/perspective.el#L969-L973

I'm not sure how current-buffer and window visibility is affected in the case of `vc-git-*`, though. Temporary buffers aren't supposed to go into prev and next buffer lists for a window (or if they do, that's pretty surprising behavior IMO — users aren't supposed to see them).