@mehw 2022-07-06
Hi @gcv, I see you removed `persp-feature-flag-prevent-killing-last-buffer-in-perspective` with commit https://github.com/nex3/perspective-el/commit/e994fb3067d343732f9fc0ae209cecd5a6192237.

I plan to formalize the remaining tests this week end.  I have also to fix a couple of subtle bugs ;)