@gcv 2022-07-09
No rush and no worries! We're all volunteers here, and this is about Emacs. It's all in good fun.